requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

dependency_injector == 4.40.0
python_dateutil == 2.8.2
python_dotenv == 0.21.0
requests == 2.28.1
